Repackage core and infrastructure met modules
