+++ core/met/met_api/src/main/java/com/vrickey123/met_api/MetObject.kt
@@ -1,0 +1,25 @@
+package com.vrickey123.met_api
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity
+data class MetObject(
+    @PrimaryKey
+    val objectID: Int,
+    val isHighlight: Boolean,
+    val isPublicDomain: Boolean,
+    val primaryImage: String,
+    val primaryImageSmall: String,
+    val department: String,
+    val objectName: String,
+    val title: String,
+    val artistDisplayName: String,
+    val artistDisplayBio: String,
+    val artistNationality: String,
+    val objectDate: String,
+    val medium: String,
+    val dimensions: String,
+    val objectURL: String,
+    val GalleryNumber: String
+)
