+++ core/met/met_api/src/main/java/com/vrickey123/met_api/MetSearchResult.kt
@@ -1,0 +1,6 @@
+package com.vrickey123.met_api
+
+data class MetSearchResult(
+    val total: Int,
+    val objectIDs: List<Int>
+)
