--- core/model/src/main/java/com/vrickey123/model/api/MetSearchResult.kt
@@ -1,6 +1,0 @@
-package com.vrickey123.model.api
-
-data class MetSearchResult(
-    val total: Int,
-    val objectIDs: List<Int>
-)
