+++ core/state/build.gradle
@@ -28,8 +28,3 @@ android {
         jvmTarget = '1.8'
     }
 }
-
-dependencies {
-    implementation "androidx.room:room-runtime:$room_version"
-    annotationProcessor "androidx.room:room-compiler:$room_version"
-}
