+++ feature/painting/build.gradle
@@ -38,14 +38,15 @@ android {
 }
 
 dependencies {
-    implementation project(':infrastructure:network')
+    implementation project(':infrastructure:met:met_network')
     implementation project(':infrastructure:reducer')
     implementation project(':infrastructure:router')
     implementation project(':infrastructure:screen')
-    implementation project(':infrastructure:ui_component')
+    implementation project(':infrastructure:met:met_ui_component')
     implementation project(':infrastructure:viewmodel')
     implementation project(':infrastructure:image')
-    implementation project(':core:model')
+    implementation project(':core:met:met_api')
+    implementation project(':core:state')
 
     implementation "androidx.core:core-ktx:$core_ktx_version"
     implementation "androidx.compose.ui:ui:$compose_version"
