+++ infrastructure/met/met_network/src/main/AndroidManifest.xml
