+++ infrastructure/met/met_network/src/main/java/com/vrickey123/network/local/MetObjectDao.kt
@@ -1,0 +1,47 @@
+package com.vrickey123.network.local
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import com.vrickey123.met_api.MetObject
+import kotlinx.coroutines.flow.Flow
+
+@Dao
+interface MetObjectDAO {
+    @Query("SELECT * FROM metobject")
+    fun loadAllAsFlow(): Flow<List<com.vrickey123.met_api.MetObject>>
+
+    @Query("SELECT * FROM metobject")
+    suspend fun loadAll(): List<com.vrickey123.met_api.MetObject>
+
+    @Query("SELECT * FROM metobject WHERE objectID = :id")
+    fun loadByIdAsFlow(id: Int): Flow<com.vrickey123.met_api.MetObject>
+
+    @Query("SELECT * FROM metobject WHERE objectID = :id")
+    fun loadByIdsAsFlow(id: List<Int>): Flow<List<com.vrickey123.met_api.MetObject>>
+
+    @Query("SELECT * FROM metobject WHERE objectID = :id")
+    fun loadAsFlow(id: Int): Flow<com.vrickey123.met_api.MetObject>
+
+    @Query("SELECT * FROM metobject WHERE objectID = :id")
+    suspend fun load(id: String): com.vrickey123.met_api.MetObject
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insertList(metObjects: List<com.vrickey123.met_api.MetObject>)
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insert(metObject: com.vrickey123.met_api.MetObject)
+
+    @Query("DELETE FROM metObject")
+    suspend fun deleteAll()
+
+    @Query("DELETE FROM metObject WHERE objectID = :id")
+    suspend fun deleteById(id: Int)
+
+    @Query("DELETE FROM metObject WHERE objectID = :ids")
+    suspend fun deleteByIds(ids: List<Int>)
+
+    @Query("SELECT (SELECT COUNT(*) FROM metObject) == 0")
+    suspend fun isEmpty(): Boolean
+}
