+++ infrastructure/met/met_ui_component/src/main/AndroidManifest.xml
