+++ settings.gradle
@@ -14,15 +14,15 @@ dependencyResolutionManagement {
 }
 rootProject.name = "The Met Impressionist Showcase"
 include ':app'
-include ':infrastructure:network'
-include ':core:model'
+include ':feature:showcase'
+include ':feature:painting'
 include ':infrastructure:viewmodel'
 include ':infrastructure:reducer'
 include ':infrastructure:screen'
 include ':infrastructure:router'
-include ':infrastructure:ui_component'
-include ':feature:showcase'
-include ':feature:painting'
 include ':infrastructure:image'
-include ':infrastructure:met_route'
+include ':infrastructure:met:met_route'
+include ':infrastructure:met:met_network'
+include ':infrastructure:met:met_ui_component'
 include ':core:state'
+include ':core:met:met_api'
